requirements.txt
scipy
numpy
matplotlib.pyplot
matplotlib.widgets
collections
sklearn.cluster